refactor(utils): added type definitions for the utils module
