--- src/types.ts
@@ -1,6 +1,0 @@
-export type OptionsType = {
-  lowercase?: boolean
-  uppercase?: boolean
-  numbers?: boolean
-  symbols?: boolean
-}
